test(operator): hoist duplicated reconcile-test helpers into shared/mod.rs
Several integration-test helpers were copy-pasted across the operator's
OAuth/listener reconcile test files. Hoist them into shared/mod.rs as
`pub fn` and import from there, removing ~218 lines of duplication.

Hoisted:
- rule_get_secret / rule_get_secret_404
- rules_for_failure_path
- assert_ready_false_with_reason
- extract_broker0_toml (canonical = the richer "data keys = [...]"
  diagnostic variant from the listener_oauth/listener_auth files)
- pool_cr (parent: &str) and pool_reconcile_rules (4-arg)

pool_ctx was byte-identical to the existing shared build_ctx, so its
call sites now use build_ctx instead of adding a redundant twin. The
same-named-but-different pool_cr (Option<&str>) in reconcile_pool.rs and
pool_reconcile_rules (3-arg) in reconcile_inter_broker_mtls.rs are
genuinely distinct and left untouched.

Test-only cleanup, behavior identical. Verified: cargo test -p
crabka-operator (all pass), cargo clippy -p crabka-operator --tests
(clean), cargo fmt --check (clean).
