+++ crates/operator/tests/reconcile_oauth_trust.rs
@@ -9,24 +9,24 @@
 //! integration tests exercise the Secret-touching code paths and the
 //! status-condition wiring end-to-end.
 
-use std::collections::BTreeMap;
 use std::sync::Arc;
 
 use base64::Engine as _;
 use crabka_operator::controller::kafka::reconcile as reconcile_kafka;
 use crabka_operator::controller::kafka_node_pool::reconcile as reconcile_pool;
 use crabka_operator::crd::{
-    Kafka, KafkaNodePool, KafkaNodePoolSpec, KafkaSpec, Listener, ListenerAuthentication,
-    ListenerAuthenticationOAuth, ListenerType, NodeRole, TlsTrustedCertificate,
+    Kafka, KafkaSpec, Listener, ListenerAuthentication, ListenerAuthenticationOAuth, ListenerType,
+    TlsTrustedCertificate,
 };
-use http::{Method, Response};
+use http::Method;
 
 #[path = "shared/mod.rs"]
 mod shared;
 
 use shared::{
-    MockRule, MockState, build_ctx, fake_kafka_body, fake_pool_body, fake_pool_list_item,
-    fake_sts_body, fixture_ctx, happy_path_rules, json_response, mock_client, not_found_body,
+    MockRule, assert_ready_false_with_reason, build_ctx, fake_pool_list_item, happy_path_rules,
+    json_response, pool_cr, pool_reconcile_rules, rule_get_secret, rule_get_secret_404,
+    rules_for_failure_path,
 };
 
 // ── fixtures ────────────────────────────────────────────────────────────────
@@ -141,29 +141,6 @@ fn managed_secret_body(name: &str, namespace: &str) -> serde_json::Value {
     })
 }
 
-/// Build a rule for `GET /secrets/<name>` returning a 404. Used by the
-/// `MissingOauthTrustSecret` test.
-fn rule_get_secret_404(name: &str) -> MockRule {
-    MockRule {
-        method: Method::GET,
-        path_substr: format!("/secrets/{name}"),
-        response: Response::builder()
-            .status(404)
-            .header("content-type", "application/json")
-            .body(not_found_body("not found"))
-            .expect("404 builds"),
-    }
-}
-
-/// Build a rule for `GET /secrets/<name>` returning the supplied body.
-fn rule_get_secret(name: &str, body: &serde_json::Value) -> MockRule {
-    MockRule {
-        method: Method::GET,
-        path_substr: format!("/secrets/{name}"),
-        response: json_response(200, body),
-    }
-}
-
 /// Build a rule for `PATCH /secrets/<name>` returning a managed-Secret body.
 fn rule_patch_managed_secret(name: &str, namespace: &str) -> MockRule {
     MockRule {
@@ -171,82 +148,6 @@ fn rule_patch_managed_secret(name: &str, namespace: &str) -> MockRule {
         path_substr: format!("/secrets/{name}"),
         response: json_response(200, &managed_secret_body(name, namespace)),
     }
-}
-
-/// Trim down `happy_path_rules` for the failure-path tests: when the
-/// trust-bundle assembly fails, the reconciler short-circuits to a
-/// status PATCH on the `Kafka` CR before reading the pool list or
-/// upserting per-broker objects. Drop those rules so an unconsumed
-/// rule doesn't mask a missing trust-bundle assertion.
-///
-/// The retained rules cover: headless Service, cluster-id Secret
-/// create, cluster-CA + clients-CA pair creates, and a status PATCH
-/// rule (the trust-bundle failure surfaces `Ready=False`, which
-/// `patch_status_with_condition` fetches+patches via two requests).
-fn rules_for_failure_path(name: &str, namespace: &str) -> Vec<MockRule> {
-    let mut rules = happy_path_rules(name, namespace, &[]);
-    // Drop the rules that fire *after* the trust-bundle assembly:
-    // broker keystore, broker-config CM, and the final aggregated
-    // status PATCH (replaced by the GET + PATCH pair below for the
-    // failure-mode status surface).
-    //
-    // The pool LIST (`/namespaces/<ns>/kafkanodepools`) DOES fire
-    // before the trust-bundle assembly — keep it. Per-pool PATCHes
-    // (`/kafkanodepools/<pool>?`) only fire on the success path; the
-    // empty pool-items list in `happy_path_rules(name, ns, &[])` means
-    // no per-pool rules were added anyway, so no extra filtering
-    // needed for them.
-    rules.retain(|r| {
-        !r.path_substr.contains("-kafka-brokers")
-            && !r.path_substr.contains("/configmaps/")
-            && !r.path_substr.contains(&format!("/kafkas/{name}/status"))
-    });
-    // The trust-bundle failure path calls
-    // `patch_status_with_condition`, which does GET status + PATCH
-    // status. Add a rule for each.
-    rules.push(MockRule {
-        method: Method::GET,
-        path_substr: format!("/kafkas/{name}/status"),
-        response: json_response(200, &fake_kafka_body(name, namespace)),
-    });
-    rules.push(MockRule {
-        method: Method::PATCH,
-        path_substr: format!("/kafkas/{name}/status"),
-        response: json_response(200, &fake_kafka_body(name, namespace)),
-    });
-    rules
-}
-
-/// Find the `Ready=False` condition in the status PATCH body and
-/// assert its `reason` matches. The trust-bundle failure path patches
-/// the `Ready` condition (not `ListenersValid` — listener validation
-/// already succeeded; the failure is in the Secret-touching code that
-/// runs *after* validation).
-fn assert_ready_false_with_reason(
-    observed: &[http::Request<hyper::body::Bytes>],
-    cluster: &str,
-    expected_reason: &str,
-) {
-    let status_patch = observed
-        .iter()
-        .find(|r| {
-            r.method() == Method::PATCH
-                && r.uri()
-                    .to_string()
-                    .contains(&format!("/kafkas/{cluster}/status"))
-        })
-        .expect("status PATCH captured");
-    let body: serde_json::Value =
-        serde_json::from_slice(status_patch.body()).expect("status body is JSON");
-    let conds = body["status"]["conditions"]
-        .as_array()
-        .expect("conditions array");
-    let ready = conds
-        .iter()
-        .find(|c| c["type"] == "Ready")
-        .unwrap_or_else(|| panic!("Ready condition present; body = {body}"));
-    assert_eq!(ready["status"], "False", "body = {body}");
-    assert_eq!(ready["reason"], expected_reason, "body = {body}");
 }
 
 /// Find the managed-Secret PATCH body and decode its `data.ca.crt` key.
@@ -470,27 +371,6 @@ async fn oauth_trust_managed_secret_updates_when_source_changes() {
 
 // ── pool-reconcile fixtures (tests 7 + 8) ──────────────────────────────────
 
-fn pool_cr(name: &str, namespace: &str, parent: &str, replicas: i32) -> KafkaNodePool {
-    let mut p = KafkaNodePool::new(
-        name,
-        KafkaNodePoolSpec {
-            roles: vec![NodeRole::Controller, NodeRole::Broker],
-            replicas,
-            node_id_start: 0,
-            image: None,
-            resources: None,
-            template: None,
-            storage: None,
-        },
-    );
-    p.metadata.namespace = Some(namespace.into());
-    p.metadata.uid = Some("pool-uid".into());
-    let mut labels = BTreeMap::new();
-    labels.insert("crabka.io/cluster".into(), parent.into());
-    p.metadata.labels = Some(labels);
-    p
-}
-
 /// Parent-Kafka body that carries an OAuth listener with
 /// `tls_trusted_certificates` populated. Used as the GET response for
 /// the pool reconciler's `kafka_api.get_opt(parent_name)` step so the
@@ -534,61 +414,6 @@ fn parent_kafka_body_no_listeners(name: &str, namespace: &str) -> serde_json::Va
     })
 }
 
-/// Build the FIFO rule sequence the pool reconciler needs:
-///   1. GET kafkas/<parent>            → parent_body
-///   2. GET statefulsets/<parent>-<pool> → 404 (first reconcile)
-///   3. PATCH statefulsets/<parent>-<pool> (SSA)
-///   4. GET statefulsets/<parent>-<pool> (post-apply status read)
-///   5. PATCH kafkanodepools/<pool>/status
-fn pool_reconcile_rules(
-    parent: &str,
-    pool: &str,
-    namespace: &str,
-    parent_body: &serde_json::Value,
-) -> Vec<MockRule> {
-    let sts_name = format!("{parent}-{pool}");
-    vec![
-        MockRule {
-            method: Method::GET,
-            path_substr: format!("/kafkas/{parent}"),
-            response: json_response(200, parent_body),
-        },
-        MockRule {
-            method: Method::GET,
-            path_substr: format!("/statefulsets/{sts_name}"),
-            response: Response::builder()
-                .status(404)
-                .header("content-type", "application/json")
-                .body(not_found_body("first reconcile, no live STS"))
-                .expect("404 builds"),
-        },
-        MockRule {
-            method: Method::PATCH,
-            path_substr: format!("/statefulsets/{sts_name}"),
-            response: json_response(200, &fake_sts_body(&sts_name, namespace, 1, Some(1))),
-        },
-        MockRule {
-            method: Method::GET,
-            path_substr: format!("/statefulsets/{sts_name}"),
-            response: json_response(200, &fake_sts_body(&sts_name, namespace, 1, Some(1))),
-        },
-        MockRule {
-            method: Method::PATCH,
-            path_substr: format!("/kafkanodepools/{pool}/status"),
-            response: json_response(200, &fake_pool_body(pool, namespace, parent)),
-        },
-    ]
-}
-
-fn pool_ctx(
-    namespace: &str,
-    rules: Vec<MockRule>,
-) -> (Arc<crabka_operator::context::Context>, Arc<MockState>) {
-    let state = MockState::new(rules);
-    let client = mock_client(&state, namespace);
-    (Arc::new(fixture_ctx(client, namespace)), state)
-}
-
 // ── test 7: StatefulSet mounts the managed trust Secret when present ───────
 
 /// Full pool reconcile with a parent Kafka CR that carries an OAuth
@@ -604,7 +429,7 @@ async fn statefulset_mounts_oauth_jwks_trust_secret_when_trust_certs_present() {
         "n7",
         &parent_kafka_body_with_oauth_trust("c7", "n7"),
     );
-    let (ctx, state) = pool_ctx("n7", rules);
+    let (ctx, state) = build_ctx("n7", rules);
     let pool = pool_cr("brokers", "n7", "c7", 1);
     reconcile_pool(Arc::new(pool), ctx).await.unwrap();
 
@@ -670,7 +495,7 @@ async fn statefulset_omits_oauth_jwks_trust_volume_when_no_trust_certs() {
         "n8",
         &parent_kafka_body_no_listeners("c8", "n8"),
     );
-    let (ctx, state) = pool_ctx("n8", rules);
+    let (ctx, state) = build_ctx("n8", rules);
     let pool = pool_cr("brokers", "n8", "c8", 1);
     reconcile_pool(Arc::new(pool), ctx).await.unwrap();
